+++ js/describe-photo.js
@@ -11,4 +11,9 @@ const getDescPhoto = (index) => {
   };
 };
 
-export {getDescPhoto}
+//массив со всеми объектами
+const describePhotos = (count) => {
+  return Array(count).fill(null).map((_, index) => getDescPhoto(index));
+}
+
+export {describePhotos}
