+++ js/main.js
@@ -1,8 +1,3 @@
-import {getDescPhoto} from './describe-photo.js';
-
-//массив со всеми объектами
-const describePhotos = (count) => {
-  return Array(count).fill(null).map((_, index) => getDescPhoto(index));
-}
+import {describePhotos} from './describe-photo.js';
 
 describePhotos(25);
